+++ lib/rubocop.rb
@@ -82,6 +82,7 @@ require 'rubocop/cop/assignment_in_condition'
 require 'rubocop/cop/end_alignment'
 require 'rubocop/cop/void'
 require 'rubocop/cop/documentation'
+require 'rubocop/cop/unreachable_code'
 
 require 'rubocop/cop/rails/validation'